requirement.txt
tqdm
addict
easydl
einops
h5py
numpy
pandas
Pillow
pyarrow
pyparsing
PyYAML
scanpy
scikit-learn
scipy
scprep
tifffile==2022.10.10
torchmetrics==0.10.2
pytorch-lightning==1.8.0.post1
torchvision
